Check path for deb-to-kali.

+++ deb-to-kali.sh
@@ -6,6 +6,10 @@ if [ `whoami` != "root" ] ; then
 fi
 
 BASEDIR=`dirname $0`
+if ! -f ${BASEDIR}/keys/gpg/kali-repo.key ; then
+  echo "Couldn't find key, are you in the right place?" >&2
+  exit 1
+fi
 
 cat >/etc/apt/sources.list.d/kali.list <<KALI_EOF
 deb http://http.kali.org/kali kali-rolling main contrib non-free
